Allow pygmy to observe changes to the cluster since the last run.
Pygmy cannot assume it will be the only thing making changes to
the AWS environment, and running get_all_db_data all the time is
hardly an elegant solution. Instead, at the start of each rule,
observe the state of AWS and compare that to pygmy's database.

Also fix some logging along the way.