Allow for rule conditions to be OR'd together.
This lets us upsize a cluster if load is too high
OR if replication falls behind.

Zero work went into allowing for the UI to make use
of this, but the API lets you do so. When making a
rule from the UI, the default logic should be AND'd
conditions.