First pass at making pygmy crash-proof.
In case pygmy crashes while resizing an instance, add the concept
of logging the intent of which instances was being shut down
to change its size. Then, if we die while the instance is stopped,
we can attempt to start the instance upon reboot, and retry the
rule.

If we die at some other place in the rule, we won't know about our
intent to run the rule, but it seems like the downsizes there are
1. we don't downsize and we waste money for a day, or
2. we don't upsize. We should probably address this in the future.