Allow Pygmy to find DNS entries based on tag as well as ID.
Restricting Pygmy to only be able to determine DNS entries based
on initial configuration (i.e. always use cname X when working on
instance Y) let to fragile configurations that struggled after a
primary failover, or when nodes were added or removed to the cluster.

This change lets us look up the CNAME to manipulate based on either
the instance, as we had before, or also by tag.

One janky thing I don't like about this, but do not know how to fix,
is that rest_framework seems to be unhappy with nullable foreign keys
being optional. So when using the API, it becomes necessary to
mention fields you don't care about and set their values to null
explicitly, rather than just leave them out of the PUT call. That's
not the worst thing but it would be nicer to just leave them out.