Handle upsize load checks correctly.
When our rule specifies a load-triggered upsize, we need to check
the load on the primary regardless of what else might be going on
with the replicas. Before this patch, if the lag or connections
on the replicas wouldn't trigger the rule, we never would check
load.

With this patch, if load triggers an upscale, we won't bother
testing the individual replicas, because we know we're going
to upsize them anyway.